+++ apps/mobile/src/components/transactions-list/Card/TxContractInteractionCard/TxContractInteractionCard.test.tsx
@@ -8,6 +8,7 @@ describe('TxContractInteractionCard', () => {
   it('should render the default markup', () => {
     const { getByText, getByLabelText } = render(
       <TxContractInteractionCard
+        onPress={() => null}
         txInfo={
           mockTransferWithInfo({
             type: TransactionInfoType.CUSTOM,
@@ -28,6 +29,7 @@ describe('TxContractInteractionCard', () => {
   it('should render a fallback in the label and icon if the contract is missing name and logoUri', () => {
     const { getByText } = render(
       <TxContractInteractionCard
+        onPress={() => null}
         txInfo={
           mockTransferWithInfo({
             type: TransactionInfoType.CUSTOM,
