+++ apps/mobile/src/components/transactions-list/Card/TxRejectionCard/TxRejectionCard.test.tsx
@@ -5,7 +5,9 @@ import { Cancellation } from '@safe-global/store/gateway/types'
 
 describe('TxRejectionCard', () => {
   it('should render the default markup', () => {
-    const { getByText } = render(<TxRejectionCard txInfo={mockTransferWithInfo({}) as Cancellation} />)
+    const { getByText } = render(
+      <TxRejectionCard onPress={() => null} txInfo={mockTransferWithInfo({}) as Cancellation} />,
+    )
 
     expect(getByText('Rejected')).toBeTruthy()
   })
