+++ apps/mobile/src/components/transactions-list/Card/TxSwapCard/TxSwapCard.test.tsx
@@ -5,7 +5,7 @@ import { mockSwapTransfer } from '@/src/tests/mocks'
 
 describe('TxSwapCard', () => {
   it('should render the default markup', () => {
-    const container = render(<TxSwapCard txInfo={mockSwapTransfer as OrderTransactionInfo} />)
+    const container = render(<TxSwapCard onPress={() => null} txInfo={mockSwapTransfer as OrderTransactionInfo} />)
 
     expect(container).toMatchSnapshot()
   })
