--- apps/mobile/src/features/ConflictTransactionSheet/index.ts
@@ -1,1 +1,0 @@
-export { ConflictTransactionSheet } from './ConflictTransactionSheet'
