+++ apps/mobile/src/features/ConflictTxSheet/index.ts
@@ -1,0 +1,1 @@
+export { ConflictTxSheetContainer } from './ConflictTxSheet.container'
